Refactor: Persist modal state in URL query parameters
This commit introduces changes to persist the state of the \"Call Wealthy\" and \"Journal\" modals, as well as the login modal, in the URL query parameters. This allows users to share links with specific modals open, maintaining the application state. Additionally, the active tab in the \"Call Wealthy\" modal is also persisted in the URL. This is achieved by using `useSearchParams` and updating the query parameters whenever the modal or tab state changes. An eslint disable comment was added to disable the exhaustive deps warning.